# Improve Songs/Artists/Albums query performances by ~50