@quentinguidee 2024-02-26
Also, I think a lot of requests can beneficiate from this improvement. If everything looks fine I might try to improve the others.