@quentinguidee 2024-02-26
I just realised that the request is sorted by listening count, not total duration, so we can directly limit/offset at start, which would make the request instantaneous