@Yooooomi 2024-02-26
Hello! Thanks a lot for what you've done. It seems good to me.

The facet is a bit weird since I'm not used to it.
Would be interesting to test if it outputs the exact same information.
This performance boost is insane thanks a lot! I don't want to abuse or anything but yeah it would be very good to have a brand new look on other queries too.
Maybe (surely) `getBestArtistsNbOffseted` and `getBestAlbumsNbOffseted` would benefit from this too.

I've tried working on #162 at some point but did not find any better implementation back then.

Again, thanks a lot. Don't hesitate if you have any questions, and keep in touch when you feel like you're done with this :)