@Yooooomi 2024-02-26
> I'll try to improve the PR in the next days to cover more improvements 👍

Thanks a lot