@quentinguidee 2024-02-27
With the same technique, https://github.com/Yooooomi/your_spotify/pull/348/commits/5669865099a6981c83c03c2d88bc878b0b56f6f6 makes the /top/artists request faster, from 12.1s to 1.3s

For this request however:
- I removed track and album from the result. I don't think it makes sense since an artists doesn't have only one track/one album, so it seems safe to remove;
- I also removed the `differents` field, it seems to come from other requests like `differentArtistsPer`.