@RomainNeup 2024-02-27
Hello @quentinguidee, congrats for the optimization it's pretty impressive 👏
If you have time, could you please have a look at the new mongoDb requests added in my PR https://github.com/Yooooomi/your_spotify/pull/346 to see if it can also be optimized ?