@quentinguidee 2024-02-27
@Yooooomi I have a MongoDB question. I think that almost every requests in this project could be instantaneous if we could have `duration_ms`, `track_id`, `album_id`, `artist_id` directly indexed in the InfosModel. I don't see another way to do this than adding these fields to all InfosModel. Do you know any way to achieve this automatically with MongoDB?