@Yooooomi 2024-02-27
I can write a migration that adds these fields in the info model.
I'll be working on it later today if this can help you.