@quentinguidee 2024-02-27
Thanks! That would be helpful. Also this might open the path for more precise total duration when storing each listening duration separately