@Yooooomi 2024-02-27
Spotify history is sometimes not precise, relying on timestamps to establish listening duration might be dangerous.
I'm pushing the migration to release/1.8.0 branch. You will be able to rebase/merge from there to have access to albumId, artistId and durationMs in the Infos model.