@Yooooomi 2024-02-27
It did trigger on my instance. Are you sure you have the latest code built?