@quentinguidee 2024-02-27
Oh the migration was marked as done in the table "migrations", but was not done for a reason I don't understand. I did a mongorestore and restarted, and it worked