@Yooooomi 2024-02-27
Nice job! I don't remember why we compute total_duration_ms. This feels weird when we just take a subpart of the whole thing.