@quentinguidee 2024-02-27
I think this is for the percentages here:
<img width="477" alt="image" src="https://github.com/Yooooomi/your_spotify/assets/12123721/17f4655b-0f34-4eeb-ac67-ba0bbf8374c1">

But if we remove them the request is extremely fast