@Yooooomi 2024-02-27
I feel like we could do this in a request on the side. So that this request does not have to compute this for every page.