@quentinguidee 2024-02-28
Thanks! Everything looks fine, the 3 requests are almost instantaneous now. I have also refactor the 3 to use the same base request. I think the "differents" field and some lookup can be removed but that doesn't change the time so clean code is better imo for now while we recreate the requests