@quentinguidee 2024-02-28
I'll keep this PR restricted to these 3 methods for now so we can iterate faster, so this one is ready for review 🧪