@quentinguidee 2024-02-28
Yeah that changes entirely the feeling of the app! The tradeoff with document databases is that join requests are really bad and should be avoided. Without that the performances are extremely good! I'll work on the other queries really soon, I have a lot of free time at the moment.

For the 30MB of indexes yeah that was predictable, but I think it is nothing compared to the space it took in RAM during the giant request (but I don't have benchmarks for that).